+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.  This
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.0.0] - 2017-06-29
+## [2.0.0] - 2017-08-07
 
 - Rewrote and restructured much of Metalware to be more consistent and
   flexible.